🦄 refactor(CardsContext): to object rather than array
